strat9_kernel::ostd

Module util

Utility types and functions for OSTD

Provides common utility types including:

ID sets for CPU/task sets

Bit manipulation helpers

Other common utilities

Modules

bits

Bit manipulation utilities

Structs

IdSet

A set of IDs (e.g., CPU IDs, task IDs)

IdSetIter

Iterator over IDs in an IdSet

Enums

Error

Functions

round_down

Round down a value to the nearest multiple of align

round_up

Round up a value to the nearest multiple of align

Type Aliases

CpuSet

A CPU set for tracking which CPUs are online/active

Result

Result type alias for OSTD operations